+++ packages/react-native-the-sheet/src/bottom-sheet-keyboard-expander/index.tsx
@@ -1,0 +1,2 @@
+export * from './bottom-sheet-keyboard-expander'
+export * from './types'