+++ packages/react-native-the-sheet/src/keyboard-expander/types.ts
@@ -1,3 +1,0 @@
-export type KeyboardExpanderProps = {
-  keyboardOffset?: number
-}